+++ crates/client/block_import/src/tests/block_import_utils.rs
@@ -104,6 +104,7 @@ pub fn create_dummy_block() -> PreValidatedBlock {
         receipts: vec![],
         state_diff: StateDiff::default(),
         converted_classes: Default::default(),
+        visited_segments: None,
     }
 }
 
@@ -128,6 +129,7 @@ pub fn create_dummy_unverified_full_block() -> UnverifiedFullBlock {
         declared_classes: vec![],
         commitments: UnverifiedCommitments::default(),
         trusted_converted_classes: vec![],
+        visited_segments: None,
     }
 }
 
@@ -151,5 +153,6 @@ pub fn create_dummy_pending_block() -> PreValidatedPendingBlock {
         receipts: vec![],
         state_diff: StateDiff::default(),
         converted_classes: Default::default(),
+        visited_segments: None,
     }
 }