+++ crates/client/rpc/src/utils/mod.rs
@@ -3,7 +3,7 @@ pub(crate) mod transaction;
 use std::fmt;
 
 use crate::StarknetRpcApiError;
-pub use transaction::to_blockifier_transactions;
+pub use transaction::to_blockifier_transaction;
 
 pub fn display_internal_server_error(err: impl fmt::Display) {
     tracing::error!(target: "rpc_errors", "{:#}", err);