+++ crates/client/rpc/src/versions/user/v0_7_1/methods/read/block_hash_and_number.rs
@@ -53,6 +53,8 @@ mod tests {
                 },
                 StateDiff::default(),
                 vec![],
+                None,
+                None,
             )
             .unwrap();
 
@@ -70,6 +72,8 @@ mod tests {
                 },
                 StateDiff::default(),
                 vec![],
+                None,
+                None,
             )
             .unwrap();
 
@@ -90,6 +94,8 @@ mod tests {
                 },
                 StateDiff::default(),
                 vec![],
+                None,
+                None,
             )
             .unwrap();
 
@@ -117,6 +123,8 @@ mod tests {
                 },
                 StateDiff::default(),
                 vec![],
+                None,
+                None,
             )
             .unwrap();