+++ crates/client/rpc/src/versions/user/v0_7_1/methods/read/estimate_fee.rs
@@ -1,17 +1,14 @@
-use std::sync::Arc;
-
-use starknet_types_core::felt::Felt;
-use starknet_types_rpc::{BroadcastedTxn, FeeEstimate, SimulationFlagForEstimateFee};
-
-use mc_exec::ExecutionContext;
-use mp_block::BlockId;
-use mp_transactions::broadcasted_to_blockifier;
-
 use crate::errors::StarknetRpcApiError;
 use crate::errors::StarknetRpcResult;
 use crate::utils::ResultExt;
 use crate::versions::user::v0_7_1::methods::trace::trace_transaction::FALLBACK_TO_SEQUENCER_WHEN_VERSION_BELOW;
 use crate::Starknet;
+use mc_exec::ExecutionContext;
+use mp_block::BlockId;
+use mp_transactions::BroadcastedTransactionExt;
+use starknet_types_core::felt::Felt;
+use starknet_types_rpc::{BroadcastedTxn, FeeEstimate, SimulationFlagForEstimateFee};
+use std::sync::Arc;
 
 /// Estimate the fee associated with transaction
 ///
@@ -40,7 +37,7 @@ pub async fn estimate_fee(
 
     let transactions = request
         .into_iter()
-        .map(|tx| broadcasted_to_blockifier(tx, starknet.chain_id(), starknet_version).map(|(tx, _)| tx))
+        .map(|tx| tx.into_blockifier(starknet.chain_id(), starknet_version).map(|(tx, _)| tx))
         .collect::<Result<Vec<_>, _>>()
         .or_internal_server_error("Failed to convert BroadcastedTransaction to AccountTransaction")?;