+++ crates/client/rpc/src/versions/user/v0_7_1/methods/trace/mod.rs
@@ -1,7 +1,4 @@
-pub(crate) mod simulate_transactions;
-pub mod trace_block_transactions;
-pub(crate) mod trace_transaction;
-
+use crate::{versions::user::v0_7_1::StarknetTraceRpcApiV0_7_1Server, Starknet};
 use jsonrpsee::core::{async_trait, RpcResult};
 use mp_block::BlockId;
 use simulate_transactions::simulate_transactions;
@@ -10,7 +7,9 @@ use starknet_types_rpc::{BroadcastedTxn, SimulateTransactionsResult, SimulationF
 use trace_block_transactions::trace_block_transactions;
 use trace_transaction::trace_transaction;
 
-use crate::{versions::user::v0_7_1::StarknetTraceRpcApiV0_7_1Server, Starknet};
+pub(crate) mod simulate_transactions;
+pub mod trace_block_transactions;
+pub(crate) mod trace_transaction;
 
 #[async_trait]
 impl StarknetTraceRpcApiV0_7_1Server for Starknet {