+++ crates/node/src/cli/chain_config_overrides.rs
@@ -49,6 +49,10 @@ pub struct ChainConfigOverridesInner {
     #[serde(skip_serializing)]
     #[serde(deserialize_with = "deserialize_private_key")]
     pub private_key: ZeroingPrivateKey,
+    pub mempool_tx_limit: usize,
+    pub mempool_declare_tx_limit: usize,
+    #[serde(deserialize_with = "deserialize_duration", serialize_with = "serialize_duration")]
+    pub mempool_tx_max_age: Duration,
 }
 
 impl ChainConfigOverrideParams {
@@ -69,6 +73,9 @@ impl ChainConfigOverrideParams {
             eth_core_contract_address: chain_config.eth_core_contract_address,
             eth_gps_statement_verifier: chain_config.eth_gps_statement_verifier,
             private_key: chain_config.private_key,
+            mempool_tx_limit: chain_config.mempool_tx_limit,
+            mempool_declare_tx_limit: chain_config.mempool_declare_tx_limit,
+            mempool_tx_max_age: chain_config.mempool_tx_max_age,
             feeder_gateway_url: chain_config.feeder_gateway_url,
             gateway_url: chain_config.gateway_url,
         })
@@ -120,6 +127,9 @@ impl ChainConfigOverrideParams {
             versioned_constants,
             eth_gps_statement_verifier: chain_config_overrides.eth_gps_statement_verifier,
             private_key: chain_config_overrides.private_key,
+            mempool_tx_limit: chain_config_overrides.mempool_tx_limit,
+            mempool_declare_tx_limit: chain_config_overrides.mempool_declare_tx_limit,
+            mempool_tx_max_age: chain_config_overrides.mempool_tx_max_age,
         })
     }
 }