+++ crates/node/src/main.rs
@@ -15,12 +15,13 @@ use mc_analytics::Analytics;
 use mc_block_import::BlockImporter;
 use mc_db::{DatabaseService, TrieLogConfig};
 use mc_gateway_client::GatewayProvider;
-use mc_mempool::{GasPriceProvider, L1DataProvider, Mempool};
+use mc_mempool::{GasPriceProvider, L1DataProvider, Mempool, MempoolLimits};
 use mc_rpc::providers::{AddTransactionProvider, ForwardToProvider, MempoolAddTxProvider};
 use mc_telemetry::{SysInfo, TelemetryService};
 use mp_oracle::pragma::PragmaOracleBuilder;
 use mp_utils::service::{Service, ServiceGroup};
 use service::{BlockProductionService, GatewayService, L1SyncService, L2SyncService, RpcService};
+use std::sync::Arc;
 
 const GREET_IMPL_NAME: &str = "Madara";
 const GREET_SUPPORT_URL: &str = "https://github.com/madara-alliance/madara/issues";
@@ -125,7 +126,13 @@ async fn main() -> anyhow::Result<()> {
     let l1_data_provider: Arc<dyn L1DataProvider> = Arc::new(l1_gas_setter.clone());
 
     // declare mempool here so that it can be used to process l1->l2 messages in the l1 service
-    let mempool = Arc::new(Mempool::new(Arc::clone(db_service.backend()), Arc::clone(&l1_data_provider)));
+    let mut mempool = Mempool::new(
+        Arc::clone(db_service.backend()),
+        Arc::clone(&l1_data_provider),
+        MempoolLimits::new(&chain_config),
+    );
+    mempool.load_txs_from_db().context("Loading mempool transactions")?;
+    let mempool = Arc::new(mempool);
 
     let l1_service = L1SyncService::new(
         &run_cmd.l1_sync_params,